requirements.txt
asgiref==3.4.1
astroid==2.6.2
autopep8==1.5.7
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
fastapi==0.66.0
google-api-core==1.31.0
google-api-python-client==2.13.0
google-auth==1.33.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
greenlet==1.1.0
h11==0.12.0
httplib2==0.19.1
httptools==0.2.0
idna==3.2
isort==5.9.2
lazy-object-proxy==1.6.0
mccabe==0.6.1
packaging==21.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pydantic==1.8.2
pylint==2.9.3
pyparsing==2.4.7
python-dotenv==0.18.0
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.21
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
uvicorn==0.14.0
uvloop==0.15.3
watchgod==0.7
websockets==9.1
wrapt==1.12.1
